DCID = Hash(ClientHello)
The change mitigates the attack described by @ekr; a MOTS attacker racing
spoofed Initial packets carrying the same CIDs and different but valid
ClientHello to create a colliding connection state on the server.

In addition, we require the client to continue using the same protection keys
even when a Retry updates the server CID.  The server is expected to use the
ODCID it retains (through the use of the token) to unprotect the payload.  The
benefit of the approach is that servers will automatically drop the Initial
packets with invalid tokens due to not being able to unprotect the payload.